requirements.txt
# Bot dependencies
aiogram==3.4.1
aiohttp==3.9.1

# Database
aiomysql==0.2.0
PyMySQL==1.1.0

# AI
google-genai>=1.0.0

# Admin Panel
flask==3.0.0
flask-cors==4.0.0

# HTTP Client
httpx>=0.25.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
apscheduler==3.10.4

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
